:recycle: refactor(cleanup): 단일 컴포넌트들을 한곳에 모아놓는다
