+++ src/index.ts
@@ -1,15 +1,15 @@
 import './scss/style.scss';
-import './components/navigation/navigation';
+import './components/modules/navigation';
 import './components/product-manage/product-manage-tab';
 import './components/charge-money/charge-money-tab';
 import './components/purchase-product/purchase-product-tab';
 import './components/purchase-product/insert-money-form';
 import './components/purchase-product/product-menu';
 import './components/purchase-product/release-coin';
-import './components/login-btn/login-btn';
-import './components/logged-in-btn/logged-in-btn';
-import './components/back-arrow/back-arrow';
-import './components/side-toast/side-toast';
+import './components/modules/login-btn';
+import './components/modules/logged-in-btn';
+import './components/modules/back-arrow';
+import './components/modules/side-toast';
 import './pages/vendingmachine-page';
 import './pages/login-page';
 import './pages/register-page';
